finition du tp 6 + rectification des erreur sql du tp 5 que j'ai fais ! a la place j'ai tout mis en jpql sauf le delete des utilisateur sans animaux que j'ai mis en sql

+++ src/main/java/fr/diginamic/Spring_Data_JPA/repository/PersonRepository.java
@@ -12,7 +12,7 @@ import fr.diginamic.Spring_Data_JPA.model.Animal;
 import fr.diginamic.Spring_Data_JPA.model.Person;
 
 @Repository
-public interface PersonRepository extends CrudRepository<Person, Long>, PersonRepositoryCustom {
+public interface PersonRepository extends CrudRepository<Person, Integer>, PersonRepositoryCustom {
     List<Person> findPeopleDistinctByLastnameOrFirstname(String lastname, String firstname);
 
     List<Person> findByAgeGreaterThanEqual(int age);
@@ -22,5 +22,6 @@ public interface PersonRepository extends CrudRepository<Person, Long>, PersonRe
 
     @Query("from Person p  where :Animal member of p.animals")
     List<Person> findAllPersonByAnimal(@Param("Animal") Animal animals);
+    
 
 }
